Use CONCAT on backend to gather user name; fix tests and their names
This commit fixes the concat on the backend; this is used to gather there the name of the user rather than combining it from the JS side.
Then, the tests for the GetUsersWithAnnotations are fixed accordingly, as well as solving a bug caused by their conflicting name.
