+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/assets/scripts/main.js": "/assets/scripts/main.js?id=aab5b44cb53ea4ac5410eef6aab7f437",
+    "/assets/scripts/main.js": "/assets/scripts/main.js?id=4026c957cab6938537ddafbd3b6db238",
     "/assets/styles/main.css": "/assets/styles/main.css?id=d4f68d6250d8c96e4a40e4aa7e03f4ec"
 }
