+++ tests/Http/Controllers/Api/Projects/GetUsersWithAnnotationsProjectTest.php
@@ -3,26 +3,27 @@
 namespace Biigle\Tests\Modules\Largo\Http\Controllers\Api\Projects;
 
 use ApiTestCase;
-use Biigle\MediaType;
 use Biigle\Tests\ImageAnnotationLabelTest;
 use Biigle\Tests\ImageAnnotationTest;
 use Biigle\Tests\ImageTest;
-use Biigle\Tests\UserTest;
 use Biigle\Tests\VideoAnnotationLabelTest;
 use Biigle\Tests\VideoAnnotationTest;
 use Biigle\Tests\VideoTest;
 use Biigle\Tests\VolumeTest;
+use Biigle\Tests\UserTest;
+use Biigle\MediaType;
 
-class GetUsersWithAnnotationsTest extends ApiTestCase
+class GetUsersWithAnnotationsProjectTest extends ApiTestCase
 {
-    public function testGetUsersWithAnnotations()
+    public function testGetUsersWithAnnotations(): void
     {
         // make sure the label tree and label are set up
         $this->labelRoot();
+        $this->user();
         $user2 = UserTest::create();
         $user3 = UserTest::create();
 
-        $image = ImageTest::create(['volume_id' => $this->volume()->id]);
+        $image = ImageTest::create(['volume_id' => $this->volume(['media_type_id' => MediaType::imageId()])->id]);
         $imageAnnotation = ImageAnnotationTest::create(['image_id' => $image->id]);
 
         $imageAnnotationLabel = ImageAnnotationLabelTest::create([
@@ -38,20 +39,20 @@ class GetUsersWithAnnotationsTest extends ApiTestCase
         $video = VideoTest::create(['volume_id' => $videoVolume->id]);
         $videoAnnotation = VideoAnnotationTest::create(['video_id' => $video->id]);
 
-        $this->videoAnnotationLabel = VideoAnnotationLabelTest::create([
+        $videoAnnotationLabel = VideoAnnotationLabelTest::create([
             'annotation_id' => $videoAnnotation->id,
             'user_id' => $user2->id,
         ]);
 
         $this->beGlobalGuest();
-        $this->get('api/v1/volumes/'.$this->volume()->id.'/users-with-annotations')->assertStatus(403);
+        $this->get('api/v1/projects/'.$this->project()->id.'/users-with-annotations')->assertStatus(403);
 
         $this->beEditor();
 
-        $expected = [['user_id' => $this->user()->id, 'firstname' => $this->user()->firstname, 'lastname' => $this->user()->lastname]];
-        $this->get('api/v1/volumes/'.$this->volume()->id.'/users-with-annotations')->assertStatus(200)->assertExactJson($expected);
-
-        $expected = [['user_id' => $user2->id, 'firstname' => $user2->firstname, 'lastname' => $user2->lastname]];
-        $this->get('api/v1/volumes/'.$videoVolume->id.'/users-with-annotations')->assertStatus(200)->assertExactJson($expected);
+        $expected = [
+            ['user_id' => $this->user()->id, 'name' => "{$this->user()->firstname} {$this->user()->lastname}"],
+            ['user_id' => $user2->id, 'name' => "{$user2->firstname} {$user2->lastname}"]
+        ];
+        $this->get('api/v1/projects/'.$this->project()->id.'/users-with-annotations')->assertStatus(200)->assertExactJson($expected);
     }
 }
